Added full support for the "Modify Spell" effect.
Reworked buffs & debuffs to use this new effect.

Added Benediction & Improved Judgement Talents

Small code cleanup

+++ cs/RetSim/RetSim/Units/Player/State/SpellState.cs
@@ -63,7 +63,7 @@ public class SpellState
 
 public class EffectBonus
 {
-    public int Flat { get; set; }
+    public float Flat { get; set; }
     public float Percent { get; set; }
 
     public EffectBonus()
